bundle files with webpack
