+++ package.json
@@ -1,0 +1,29 @@
+{
+    "name": "geo-quiz",
+    "version": "1.0.0",
+    "description": "Web-based game guessing what country in map built using D3.js",
+    "private": true,
+    "scripts": {
+        "build": "webpack --config webpack.config.js"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/Frankuy/geo-quiz.git"
+    },
+    "author": "Muhammad Fikri Hizbullah",
+    "license": "MIT",
+    "bugs": {
+        "url": "https://github.com/Frankuy/geo-quiz/issues"
+    },
+    "homepage": "https://github.com/Frankuy/geo-quiz#readme",
+    "devDependencies": {
+        "css-loader": "^6.0.0",
+        "html-webpack-plugin": "^5.3.2",
+        "style-loader": "^3.1.0",
+        "webpack": "^5.45.1",
+        "webpack-cli": "^4.7.2"
+    },
+    "dependencies": {
+        "d3": "^7.0.0"
+    }
+}
