+++ src/index.html
@@ -1,12 +1,13 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Geo Quiz</title>
-    <link rel="stylesheet" href="/css/style.css">
 </head>
+
 <body>
     <svg id="game-svg">
         <g id="map" />
@@ -23,9 +24,6 @@
     <div id="start-game">
         <span id="countdown"></span>
     </div>
-    <script src="/vendor/d3/dist/d3.min.js"></script>
-    <script src="/js/variable.js"></script>
-    <script src="/js/sound.js"></script>
-    <script src="/js/index.js"></script>
 </body>
+
 </html>
